ai-service/requirements-simple.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Simplified AI dependencies
openai==1.6.1
pydantic==2.5.0
structlog==23.2.0

# Remove problematic dependencies for now
# langchain==0.1.0
# langchain-openai==0.0.2
# langchain-community==0.0.10
# faiss-cpu==1.7.4
# chromadb==0.4.18
# pandas==2.1.4
# numpy==1.26.2
